requirements.txt
pycaret
Flask==2.0.2
azure-storage-blob
gunicorn